Modelangem da class Entity, iniciando a criação das interfaces de JPA

+++ src/main/java/com/example/Conta_Bancaria/application/service/AccountServiceRepository.java
@@ -1,8 +1,10 @@
 package com.example.Conta_Bancaria.application.service;
 
+import java.math.BigDecimal;
+
 public interface AccountServiceRepository {
 
-    void depositAcount(Long id, Double amount);
-    void withdrawAcount(Long id, Double amount);
+    void depositAcount(Long id, BigDecimal amount);
+    void withdrawAcount(Long id, BigDecimal amount);
 
 }
